requirements.txt
boto3>=1.34.0
botocore>=1.34.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
annotated-types==0.7.0
anyio==4.11.0
boto3==1.40.55
botocore==1.40.55
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy
cycler==0.12.1
et_xmlfile==2.0.0
filelock
fonttools==4.60.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
#huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver
langchain
langchain-aws
langchain-core
#langchain-huggingface==1.0.0
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
MarkupSafe==3.0.3
matplotlib
mpmath==1.3.0
networkx
numpy
openpyxl==3.1.5
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pillow
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.22
req==1.0.0
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn
scipy
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3
xxhash==3.6.0
zstandard==0.25.0
